==> Part 1 Data Merging.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
%matplotlib inline

## Loading and Merging the Data

order_products_prior = pd.read_csv('order_products__prior.csv') # opens the csv file
print("Rows and columns:",order_products_prior.shape)
pd.DataFrame.head(order_products_prior)

order_products_train = pd.read_csv('order_products__train.csv') # opens the csv file
print("Rows and columns:",order_products_train.shape)
pd.DataFrame.head(order_products_train)

Concatenate the two datasets

order_product=pd.concat([order_products_prior, order_products_train])
print("Rows and columns:",order_product.shape)
pd.DataFrame.head(order_product)

Merge all the other datasets to the main one

products = pd.read_csv('products.csv') # opens the csv file
print("Rows and columns:",products.shape)
pd.DataFrame.head(products)

aisles = pd.read_csv('aisles.csv') # opens the csv file
print("Rows and columns:",aisles.shape)
pd.DataFrame.head(aisles)

departments = pd.read_csv('departments.csv') # opens the csv file
print("Rows and columns:",departments.shape)
pd.DataFrame.head(departments)

products_all = pd.merge(products,aisles,on='aisle_id',how='left')

products_all = pd.merge(products_all,departments,on='department_id',how='left')
print("Rows and columns:",products_all.shape)
pd.DataFrame.head(products_all)

order_products_all = pd.merge(order_product,products_all,on='product_id',how='left')
print("Rows and columns:",order_products_all.shape)
pd.DataFrame.head(order_products_all)

orders = pd.read_csv('orders.csv') # opens the csv file
print("Rows and columns:",orders.shape)
pd.DataFrame.head(orders)

orders_users = orders[['order_id','user_id']]
print("Rows and columns:",orders_users.shape)
pd.DataFrame.head(orders_users)

order_products_all=pd.merge(order_products_all,orders_users,on='order_id',how='left')
print("Rows and columns:",order_products_all.shape)
pd.DataFrame.head(order_products_all)

order_products_all.to_csv('order_products_all.csv',sep=',')

## Creating the features for each user

### A) Order Patterns

group=orders.groupby(['user_id'],as_index=False)
min_days=pd.DataFrame(group['days_since_prior_order'].agg('min'))
pd.DataFrame.head(min_days)

test = np.array([1, np.nan, 3, 4])
min(test)

max(test)

sum(test)

med(test)

avg(test)

np.mean(test)

np.median(test)
